Secured with password and editing page implemented

+++ js/redirect.js
@@ -1,18 +1,21 @@
 var query = window.location.search;
 query = query.replace("?", "");
+const hide = document.querySelector("#board");
 
 const getDataa = async () => {
   const url = `http://localhost:1337/boards/${query}`;
   const data = await fetch(url);
   const res = await data.json();
 
-  if (localStorage.getItem("check")) {
-    `http://127.0.0.1:5500/pages/questions-read.html?${query}`;
-    localStorage.clear();
-  } else {
-    window.location.replace(
-      `http://127.0.0.1:5500/pages/password.html?${query}`
-    );
+  if (res.private) {
+    if (localStorage.getItem("check")) {
+      `http://127.0.0.1:5500/pages/questions-read.html?${query}`;
+      localStorage.clear();
+    } else {
+      window.location.replace(
+        `http://127.0.0.1:5500/pages/password.html?${query}`
+      );
+    }
   }
 };
